[SHLVS-395] moved strength level logic to SecretTextFormField
