+++ src/components/form-fields/text/secret-text-form-field.tsx
@@ -1,56 +1,69 @@
-import { FieldValues } from "react-hook-form";
+import { useEffect, useState } from "react";
+import { FieldValues, useWatch } from "react-hook-form";
 
-import { FormField, InputProps } from "@/shared/ui";
+import { PasswordStrengthLevel, checkPasswordStrength } from "@/shared/libs";
+import { FormField, InputProps, StepsProgressBar } from "@/shared/ui";
 
 import { FormFieldProps, TextFormFieldProps } from "../types";
 import { SecretTextFormItem } from "./secret-text-form-item";
 
+const strengthColorMap: Record<PasswordStrengthLevel, string> = {
+  0: "bg-red-500",
+  1: "bg-orange-500",
+  2: "bg-yellow-500",
+  3: "bg-green-500",
+};
+
 type SecretTextFormFieldProps<T extends FieldValues> = FormFieldProps<T> &
   TextFormFieldProps &
-  Omit<InputProps, "onChange">;
+  Omit<InputProps, "onChange"> & {
+    showStrengthIndicator?: boolean;
+  };
 
 export function SecretTextFormField<T extends FieldValues>({
   name,
   control,
   idPrefix,
-  // TODO: use this prop
-  // showStrengthIndicator = false,
+  showStrengthIndicator = false,
   ...props
 }: SecretTextFormFieldProps<T>) {
   const inputId = idPrefix ? `${idPrefix}-${name}` : undefined;
+  const [strengthLevel, setStrengthLevel] = useState<PasswordStrengthLevel>(0);
 
-  const FormFieldComponent = (
-    <FormField
-      control={control}
-      name={name}
-      render={({ field }) => (
-        <SecretTextFormItem
-          type="password"
-          autoFocus={props.autoFocus}
-          fieldValue={field.value}
-          onChange={(x) => field.onChange(x)}
-          inputId={inputId}
-          inForm={true}
-          {...props}
+  const passwordWatched = useWatch({
+    control: control,
+    name: name,
+  });
+
+  useEffect(() => {
+    setStrengthLevel(checkPasswordStrength(passwordWatched));
+  }, [passwordWatched]);
+
+  return (
+    <>
+      <FormField
+        control={control}
+        name={name}
+        render={({ field }) => (
+          <SecretTextFormItem
+            type="password"
+            autoFocus={props.autoFocus}
+            fieldValue={field.value}
+            onChange={(x) => field.onChange(x)}
+            inputId={inputId}
+            inForm={true}
+            {...props}
+          />
+        )}
+      />
+      {showStrengthIndicator && passwordWatched && (
+        <StepsProgressBar
+          steps={["0", "1", "2", "3"]}
+          current={strengthLevel.toString()}
+          activeColor={strengthColorMap[strengthLevel]}
+          inactiveColor="bg-transparent"
         />
       )}
-    />
+    </>
   );
-
-  return FormFieldComponent;
-
-  // NOTE: DRAFT version
-  // return showStrengthIndicator ? (
-  //   <div>
-  //     <FormField />
-  //     <StepsProgressBar
-  //       steps={["very-weak", "weak", "medium", "strong"]}
-  //       current={strength}
-  //       activeColor={activeColor}
-  //       inactiveColor={inactiveColor}
-  //     />
-  //   </div>
-  // ) : (
-  //   FormField
-  // );
 }
